backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.8.0
pillow==10.0.1
requests==2.31.0
torchvision==0.16.1
sentence-transformers==2.2.2
datasets==2.14.6
accelerate==0.24.1
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78